+++ HISTORY.md
@@ -3,6 +3,7 @@ History of features added to LiChess Tools in time (reversed order)
 28 Jul
 
 - removed the "No study chat" theme (use "Toggle study chat")
+- added Ctrl-C as general key shortcut to copy the FEN
 
 19 Jun